Add HasOption to Table Filter Implement traits in both SelectFilter and MultiSelectFilter
