+++ packages/tables/src/Filters/SelectFilter.php
@@ -12,13 +12,12 @@ use Illuminate\Support\Str;
 
 class SelectFilter extends Filter
 {
+    use Concerns\HasOptions;
     use Concerns\HasPlaceholder;
 
     protected string | Closure | null $column = null;
 
     protected bool | Closure $isStatic = false;
-
-    protected array | Arrayable | Closure | null $options = null;
 
     protected function setUp(): void
     {
@@ -62,13 +61,6 @@ class SelectFilter extends Filter
         return $this;
     }
 
-    public function options(string | array | Arrayable | Closure | null $options): static
-    {
-        $this->options = $options;
-
-        return $this;
-    }
-
     public function relationship(string $relationshipName, string $displayColumnName): static
     {
         $this->column("{$relationshipName}.{$displayColumnName}");
@@ -86,29 +78,6 @@ class SelectFilter extends Filter
     public function getColumn(): string
     {
         return $this->evaluate($this->column) ?? $this->getName();
-    }
-
-    public function getOptions(): array
-    {
-        $options = $this->evaluate($this->options);
-
-        if ($options === null) {
-            $options = $this->queriesRelationships() ? $this->getRelationshipOptions() : [];
-        }
-
-        if (is_string($options)) {
-            $options = collect($options::cases())
-                ->mapWithKeys(function ($eachEnumInstance) {
-                    return [$eachEnumInstance?->value ?? $eachEnumInstance->name => $eachEnumInstance->name];
-                })
-                ->toArray();
-        }
-
-        if ($options instanceof Arrayable) {
-            $options = $options->toArray();
-        }
-
-        return $options;
     }
 
     protected function getRelationshipOptions(): array
